refactor: rename enterprise specific hooks and callbacks to be more precise
